integration docs: Removed {!create-channel.md!} from specified docs.
Removed {!create-channel.md!} from all the integrations docs that
also include {!generate-webhook-url-basic.md!}. This removal was done
to simplify the process of integration for new Zulip users.
Updated generate-webhook-url-basic.md to replace the use of the
"Create the channel" step in the integration process and instead
hint about it here.
Fixes #32269.